SQLiteMigration + Github Pages + Logic Moved Into Libraries + Simplified Discord Bots

+++ service-destiny/src/routes/member.rs
@@ -1,8 +1,6 @@
 use super::responses::{
     APIClanInfoResponse, APIMemberTitleResponse, ClanInformation, MemberTitle, MemberTitleResponse, ReportOutput,
 };
-use crate::app;
-use crate::app::state::AppState;
 use crate::routes::responses::{APIMemberResponse, MemberResponse};
 use axum::extract::{Path, State};
 use axum::routing::get;
@@ -11,6 +9,8 @@ use levelcrush::axum;
 use levelcrush::axum::extract::Query;
 use levelcrush::server::APIResponse;
 use levelcrush::tracing;
+use lib_destiny::app;
+use lib_destiny::app::state::AppState;
 
 #[derive(serde::Serialize, serde::Deserialize)]
 pub struct ReportQueries {
@@ -71,8 +71,8 @@ async fn season_report(
         let modes = if let Some(input_modes) = report_queries.modes {
             input_modes
                 .split(',')
-                .map(|v| v.parse::<i32>().unwrap_or_default())
-                .collect::<Vec<i32>>()
+                .map(|v| v.parse::<i64>().unwrap_or_default())
+                .collect::<Vec<i64>>()
         } else {
             Vec::new()
         };
@@ -80,7 +80,7 @@ async fn season_report(
         let (task_started, report) = app::report::member::season(
             bungie_name,
             &modes,
-            season.parse::<i32>().unwrap_or_default(),
+            season.parse::<i64>().unwrap_or_default(),
             member.clan_is_network == 1,
             &mut state,
         )
@@ -116,8 +116,8 @@ async fn lifetime_report(
         let modes = if let Some(input_modes) = report_queries.modes {
             input_modes
                 .split(',')
-                .map(|v| v.parse::<i32>().unwrap_or_default())
-                .collect::<Vec<i32>>()
+                .map(|v| v.parse::<i64>().unwrap_or_default())
+                .collect::<Vec<i64>>()
         } else {
             Vec::new()
         };
